Mas mejoras en el backend de CCLH

+++ src/main/java/org/themarioga/cclh/commons/services/impl/RoomServiceImpl.java
@@ -3,8 +3,13 @@ package org.themarioga.cclh.commons.services.impl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 import org.themarioga.cclh.commons.dao.intf.RoomDao;
+import org.themarioga.cclh.commons.exceptions.room.RoomAlreadyExistsException;
+import org.themarioga.cclh.commons.exceptions.room.RoomDoesntExistsException;
+import org.themarioga.cclh.commons.exceptions.user.UserAlreadyExistsException;
+import org.themarioga.cclh.commons.exceptions.user.UserDoesntExistsException;
 import org.themarioga.cclh.commons.models.Room;
 import org.themarioga.cclh.commons.services.intf.RoomService;
 
@@ -22,13 +27,19 @@ public class RoomServiceImpl implements RoomService {
     public Room create(Room room) {
         logger.debug("Creating room: {}", room);
 
-        room.setCreationDate(new Date());
-        return roomDao.create(room);
+        try {
+            room.setCreationDate(new Date());
+            return roomDao.create(room);
+        } catch (DuplicateKeyException e) {
+            throw new RoomAlreadyExistsException();
+        }
     }
 
     @Override
     public Room update(Room room) {
         logger.debug("Updating room: {}", room);
+
+        if (findOne(room.getId()) == null) throw new RoomDoesntExistsException();
 
         return roomDao.update(room);
     }
@@ -37,12 +48,16 @@ public class RoomServiceImpl implements RoomService {
     public void delete(Room room) {
         logger.debug("Delete room: {}", room);
 
+        if (findOne(room.getId()) == null) throw new RoomDoesntExistsException();
+
         roomDao.delete(room);
     }
 
     @Override
     public void deleteById(long id) {
         logger.debug("Delete room by ID: {}", id);
+
+        if (findOne(id) == null) throw new RoomDoesntExistsException();
 
         roomDao.deleteById(id);
     }
